remove ustar fs
The initial thought was, to implement something that can read the format
and then port that to the vfs implementation, which is now a node graph.
Porting would have been possible, but a rewrite is probably faster and will
definitely make the implementation more efficient. In addition, we do not really
need a ustar fs, it was just easy to implement. We should probably go forward with
at least ext2, maybe even 3 or 4.