check span overlaps at compile-time
The list of memory spans is now checked at compile time.
The list must be sorted, the starting addresses must be disjunct and the entries must not overlap in memory.
If one of those criteria is not matched, a compile error with the name of the non-compliant spans will be emitted.